+++ src/WarHub.ArmouryModel.CliTool/Commands/CommandBase.cs
@@ -1,5 +1,7 @@
-﻿using Serilog;
+﻿using System.CommandLine;
+using Serilog;
 using Serilog.Core;
+using Serilog.Events;
 using Serilog.Sinks.SystemConsole.Themes;
 
 namespace WarHub.ArmouryModel.CliTool.Commands
@@ -8,12 +10,33 @@ namespace WarHub.ArmouryModel.CliTool.Commands
     {
         protected Logger Log { get; private set; }
 
+        public IConsole Console { get; set; }
+
         protected Logger SetupLogger(string verbosity)
         {
-            return Log = new LoggerConfiguration()
-                .MinimumLevel.Is(Program.GetLogLevel(verbosity))
-                .WriteTo.Console(theme: AnsiConsoleTheme.Code)
-                .CreateLogger();
+            var baseConfig = new LoggerConfiguration()
+                .MinimumLevel.Is(Program.GetLogLevel(verbosity));
+
+            var config = Console is SystemConsole ?
+                baseConfig.WriteTo.Console(theme: AnsiConsoleTheme.Code)
+                : baseConfig.WriteTo.Sink(new TestConsoleSink(Console));
+
+            return Log = config.CreateLogger();
+        }
+
+        private class TestConsoleSink : ILogEventSink
+        {
+            public TestConsoleSink(IConsole console)
+            {
+                Console = console;
+            }
+
+            public IConsole Console { get; }
+
+            public void Emit(LogEvent logEvent)
+            {
+                Console.Out.WriteLine(logEvent.RenderMessage());
+            }
         }
     }
 }